fix: react error with incorrect array component
